Add files via upload
# Quotes Generator Website

This is an inspirational quotes generator website built using HTML, CSS, and JavaScript. It displays random quotes at the click of a button, providing users with a source of motivation and inspiration. The site features a clean, responsive design to ensure a seamless user experience across different devices. Ideal for daily motivation or for sharing quotes on social media.